moved banner components to its folder. added component resultsBody. inside added two containers, one for filters and other for results
