+++ src/components/Banner/SearchBar/index.jsx
