+++ src/components/Banner/index.jsx
@@ -2,9 +2,9 @@ import styles from "./banner.module.css";
 import logoImage from "../../images/logo.png";
 import vectorImage from "../../images/vector.png";
 import Typography from "@mui/material/Typography";
-import { SearchBar } from "../SearchBar/index.jsx";
-import { LandingPageLink } from "../LandingPageLink/index.jsx";
-import { SearchButton } from "../SearchButton/index.jsx";
+import { SearchBar } from "./SearchBar/index.jsx";
+import { LandingPageLink } from "./LandingPageLink/index.jsx";
+import { SearchButton } from "./SearchButton/index.jsx";
 
 export function Banner({ size }) {
   let bannerHeight = 0;
