+++ src/pages/Results/index.jsx
@@ -1,10 +1,12 @@
 import styles from "./results.module.css";
 import { Banner } from "../../components/Banner/index.jsx";
+import { ResultsBody } from "../../components/ResultsBody/index.jsx";
 
 export function Results() {
   return (
     <div>
       <Banner size={"m"}></Banner>
+      <ResultsBody></ResultsBody>
     </div>
   );
 }
